alti

+++ fdf.h
@@ -6,7 +6,7 @@
 /*   By: rvalenti <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/29 16:15:30 by rvalenti          #+#    #+#             */
-/*   Updated: 2019/02/08 10:27:46 by rvalenti         ###   ########.fr       */
+/*   Updated: 2019/02/08 10:56:56 by rvalenti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,7 @@ typedef struct	s_data
 	int		y_max;
 	int		gap;
 	int		proj;
-	int		n;
+	double	n;
 	t_lst	*list;
 	int		**m;
 	t_point	p;
